refactor(api): split StorageService into FsStorageService and MinioStorageService
