+++ modules/fbs-core/api/src/main/scala/de/thm/ii/fbs/controller/CheckerConfigurationController.scala
@@ -6,6 +6,7 @@ import de.thm.ii.fbs.model._
 import de.thm.ii.fbs.services.checker.CheckerServiceFactoryService
 import de.thm.ii.fbs.services.checker.`trait`._
 import de.thm.ii.fbs.services.persistence._
+import de.thm.ii.fbs.services.persistence.storage.{FsStorageService, StorageService}
 import de.thm.ii.fbs.services.security.AuthService
 import de.thm.ii.fbs.util.JsonWrapper.jsonNodeToWrapper
 import org.springframework.beans.factory.annotation.Autowired
@@ -31,6 +32,8 @@ class CheckerConfigurationController {
   private val ccs: CheckrunnerConfigurationService = null
   @Autowired
   private val storageService: StorageService = null
+  @Autowired
+  private val fsStorageService: FsStorageService = null
   @Autowired
   private val taskService: TaskService = null
   @Autowired
@@ -214,7 +217,7 @@ class CheckerConfigurationController {
   @GetMapping(value = Array("/{cid}/tasks/{tid}/checker-configurations/{ccid}/main-file"))
   def getMainFile(@PathVariable cid: Int, @PathVariable tid: Int, @PathVariable ccid: Int,
                   req: HttpServletRequest, res: HttpServletResponse): Unit =
-    getFile(storageService.pathToMainFile)(storageFileName.MAIN_FILE, cid, tid, ccid, req, res)
+    getFile(fsStorageService.pathToMainFile)(storageFileName.MAIN_FILE, cid, tid, ccid, req, res)
 
   /**
     * Upload a the secondary file for a task configuration
@@ -249,7 +252,7 @@ class CheckerConfigurationController {
   @GetMapping(value = Array("/{cid}/tasks/{tid}/checker-configurations/{ccid}/secondary-file"))
   def getSecondaryFile(@PathVariable cid: Int, @PathVariable tid: Int, @PathVariable ccid: Int,
                        req: HttpServletRequest, res: HttpServletResponse): Unit =
-    getFile(storageService.pathToSecondaryFile)(storageFileName.SECONDARY_FILE, cid, tid, ccid, req, res)
+    getFile(fsStorageService.pathToSecondaryFile)(storageFileName.SECONDARY_FILE, cid, tid, ccid, req, res)
 
   private def uploadFile(fileName: String, postHook: CheckrunnerConfiguration => Unit)
                         (cid: Int, tid: Int, ccid: Int, file: MultipartFile, req: HttpServletRequest, res: HttpServletResponse): Unit = {
