+++ modules/fbs-core/api/src/main/scala/de/thm/ii/fbs/services/checker/SqlCheckerRemoteCheckerService.scala
@@ -6,6 +6,7 @@ import de.thm.ii.fbs.model.checker.{RunnerRequest, SqlCheckerSubmission, User}
 import de.thm.ii.fbs.model.{CheckrunnerConfiguration, SqlCheckerInformation, Task, Submission => FBSSubmission}
 import de.thm.ii.fbs.services.checker.`trait`._
 import de.thm.ii.fbs.services.persistence._
+import de.thm.ii.fbs.services.persistence.storage.StorageService
 import de.thm.ii.fbs.services.security.TokenService
 import org.apache.http.client.utils.URIBuilder
 import org.springframework.beans.factory.annotation.{Autowired, Value}
@@ -105,7 +106,7 @@ class SqlCheckerRemoteCheckerService(@Value("${services.masterRunner.insecure}")
                 }
                 if (!query.selAttributesRight.get) {
                   hints ++= "falsche Where-Attribute verwendet\n"
-                }
+                  }
                 if (!query.proAttributesRight.get) {
                   hints ++= "falsche Select-Attribute verwendet\n"
                 }
